:hammer: Extract SidePasteTitleView to separate file
